+++ docs/CHANGELOG.md
@@ -9,13 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create custom DjangoQL lookup field for Flaw.components (OSIDB-3479)
 - Collect NIST CVSSv4 in NVD collector (OSIDB-2300)
 
+### Changed
+- Record last impact increase in trackers (OSIDB-3448)
+
 ### Fixed
 - Remove duplicate results from advanced search (OSIDB-3482)
 - Collect Jira field metadata for only one issuetype for each project (OSIDB-3485)
 - parent_uuid field in Alert had wrong type in OpenAPI schema (OSIDB-3451)
+- Fix Jira Tracker collector to account for Vulnerability issue type (OSIDB-3489)
 
-### Changed
-- Record last impact increase in trackers (OSIDB-3448)
 
 ## [4.3.3] - 2024-09-30
 ### Added